refactor(phase4): extract eon_parser test suite to tests.rs
Moves the 4,905-line inline #[cfg(test)] mod tests { ... } block to
eon_parser/tests.rs, reducing eon_parser.rs from 5,098 lines to 193.

The parent module now exposes only its API surface and module wiring.
Test helpers (build_resources, parse_document, temp_root) remain
accessible because Rust child modules can read parent private items.

Baseline: 233 passing / 19 pre-existing failing (unchanged).
